Fix canvas hint debug policy findings
