+++ Packages/macOS/CmuxCanvasUI/Sources/CmuxCanvasUI/CanvasRootView+CommandScroll.swift
@@ -121,6 +121,9 @@ extension CanvasRootView {
     }
 
 #if DEBUG
+    /// Shows the Command+scroll discovery hint immediately for debug-menu and
+    /// automation testing, replacing any visible hint and leaving the
+    /// production one-time discovery flag unchanged.
     public func debugShowCommandScrollHint() {
         commandScrollHintTask?.cancel()
         presentCommandScrollHint(markSessionShown: false, replacingExisting: true)
