+++ Packages/macOS/CmuxControlSocket/Tests/CmuxControlSocketTests/ControlCommandCoordinatorDebugCanvasTests.swift
@@ -4,19 +4,6 @@ import Testing
 
 #if DEBUG
 @MainActor
-private final class FakeDebugCanvasControlCommandContext: ControlCommandContext {
-    var resolution: ControlCanvasActionResolution = .ok(mode: "canvas")
-    var lastRouting: ControlRoutingSelectors?
-
-    func controlDebugShowCanvasCommandScrollHint(
-        routing: ControlRoutingSelectors
-    ) -> ControlCanvasActionResolution {
-        lastRouting = routing
-        return resolution
-    }
-}
-
-@MainActor
 @Suite("ControlCommandCoordinator debug canvas dispatch")
 struct ControlCommandCoordinatorDebugCanvasTests {
     private func makeCoordinator() -> (ControlCommandCoordinator, FakeDebugCanvasControlCommandContext) {
